+++ packages/article_service/lib/article_service.dart
@@ -4,4 +4,3 @@
 library;
 
 export 'src/article_service_base.dart';
-export 'src/category_service_base.dart';