+++ packages/contact_service/lib/src/contact_service_base.dart
@@ -33,7 +33,7 @@ class ContactService extends ContactServiceBase {
 
     final userPermission = isTest
         ? userPermissionIfTest ?? UserPermissions()
-        : call.bearer.userPermission;
+        : call.bearer.userPermissions;
     if (userPermission.isChainAccessible(request.chainId) == false) {
       throw GrpcError.permissionDenied(
           'user cannot access data from chain ${request.chainId}');
@@ -50,14 +50,15 @@ class ContactService extends ContactServiceBase {
         ..contactId = request.contact.id
         ..chainId = request.chainId
         ..firmId = userPermission.firmId
-        ..userId = userPermission.userId;
+        ..userId = userPermission.userId
+        ..lastTouchTimestampUTC = DateTime.now().toUtc().timestampProto;
 
       final result = await collection
           .insertOne(contactMongo.toProto3Json() as Map<String, dynamic>);
       if (result.hasWriteErrors) {
         throw GrpcError.unknown('hasWriteErrors ${result.writeError!.errmsg}');
       }
-      if (result.ok == 1 && result.document != null) {
+      if (result.success && result.document != null) {
         final contactId = result.document!['contactId'] as int;
 
         return StatusResponse.create()
@@ -81,12 +82,12 @@ class ContactService extends ContactServiceBase {
   }
 
   @override
-  Future<StatusResponse> replaceOne(
+  Future<StatusResponse> updateOne(
       ServiceCall? call, ContactRequest request) async {
     _db.isConnected ? null : await _db.open();
     final userPermission = isTest
         ? userPermissionIfTest ?? UserPermissions()
-        : call.bearer.userPermission;
+        : call.bearer.userPermissions;
     if (userPermission.isChainAccessible(request.chainId) == false) {
       throw GrpcError.permissionDenied(
           'user cannot access data from chain ${request.chainId}');
@@ -103,7 +104,8 @@ class ContactService extends ContactServiceBase {
         ..contactId = request.contact.id
         ..chainId = request.chainId
         ..firmId = userPermission.firmId
-        ..userId = userPermission.userId;
+        ..userId = userPermission.userId
+        ..lastTouchTimestampUTC = DateTime.now().toUtc().timestampProto;
 
       final result = await collection.replaceOne(
           _Helpers.selectContact(
@@ -140,7 +142,7 @@ class ContactService extends ContactServiceBase {
     _db.isConnected ? null : await _db.open();
     final userPermission = isTest
         ? userPermissionIfTest ?? UserPermissions()
-        : call.bearer.userPermission;
+        : call.bearer.userPermissions;
     if (userPermission.contactRights.rights.any((e) => e == Right.delete) ==
         false) {
       throw GrpcError.permissionDenied(
@@ -174,7 +176,7 @@ class ContactService extends ContactServiceBase {
     _db.isConnected ? null : await _db.open();
     final userPermission = isTest
         ? userPermissionIfTest ?? UserPermissions()
-        : call.bearer.userPermission;
+        : call.bearer.userPermissions;
     if (userPermission.isChainAccessible(request.chainId) == false) {
       throw GrpcError.permissionDenied(
           'user cannot access data from chain ${request.chainId}');
@@ -185,13 +187,24 @@ class ContactService extends ContactServiceBase {
           'user does not have right to read contacts');
     }
     try {
-      final list = await collection.find().toList();
+      var selector = SelectorBuilder();
+      if (request.lastFetchTimestampUTC.hasSeconds()) {
+        selector = where.gte('lastTouchTimestampUTC.seconds',
+            request.lastFetchTimestampUTC.seconds);
+      }
+      final list = await collection.find(selector).toList();
+      if (list.isEmpty) {
+        return ContactsResponse.create();
+      }
+
       final contacts = <ContactPb>[];
       for (final e in list) {
         final contactMongo = ContactMongo.create()
           ..mergeFromProto3Json(e, ignoreUnknownFields: true);
+
         contacts.add(contactMongo.contact);
       }
+
       final contactsBis = ContactsResponse();
       contactsBis.contacts
         ..clear()
@@ -209,7 +222,7 @@ class ContactService extends ContactServiceBase {
     _db.isConnected ? null : await _db.open();
     final userPermission = isTest
         ? userPermissionIfTest ?? UserPermissions()
-        : call.bearer.userPermission;
+        : call.bearer.userPermissions;
     if (userPermission.isChainAccessible(request.contactChainId) == false) {
       throw GrpcError.permissionDenied(
           'user cannot access data from chain ${request.contactChainId}');
@@ -236,6 +249,68 @@ class ContactService extends ContactServiceBase {
     } on GrpcError catch (e) {
       print(e);
       rethrow;
+    }
+  }
+
+  @override
+  Future<StatusResponse> createMany(
+      ServiceCall call, ContactsRequest request) async {
+    _db.isConnected ? null : await _db.open();
+
+    final userPermission = isTest
+        ? userPermissionIfTest ?? UserPermissions()
+        : call.bearer.userPermissions;
+    if (userPermission.isChainAccessible(request.chainId) == false) {
+      throw GrpcError.permissionDenied(
+          'user cannot access data from chain ${request.chainId}');
+    }
+    if (userPermission.contactRights.rights.any((e) => e == Right.create) ==
+        false) {
+      throw GrpcError.permissionDenied(
+          'user does not have right to create articles');
+    }
+    final nowProto = DateTime.now().toUtc().timestampProto;
+    final contactsMap = <Map<String, dynamic>>[];
+    for (final contactPb in request.contacts) {
+      final contactMongo = ContactMongo.create()
+        ..contact = contactPb
+        ..creationDate = contactPb.creationDate
+        ..contactId = contactPb.id
+        ..chainId = request.chainId
+        ..firmId = userPermission.firmId
+        ..userId = userPermission.userId
+        ..lastTouchTimestampUTC = nowProto;
+
+      contactsMap.add(contactMongo.toProto3Json() as Map<String, dynamic>);
+    }
+
+    try {
+      final result = await collection.insertMany(contactsMap);
+      if (result.hasWriteErrors) {
+        final writeErrorsMessages = <String>[];
+        for (final error in result.writeErrors) {
+          writeErrorsMessages.add(error.toString());
+        }
+        throw GrpcError.unknown(
+            'hasWriteErrors ${writeErrorsMessages.join("\n")}');
+      }
+      if (result.success) {
+        return StatusResponse.create()
+          ..type = StatusResponse_Type.CREATED
+          ..timestamp = DateTime.now().timestampProto;
+      } else {
+        return StatusResponse.create()
+          ..type = StatusResponse_Type.ERROR
+          ..message = 'result.failure but no writeErrorsMessages'
+          ..timestamp = DateTime.now().timestampProto;
+      }
+    } on GrpcError catch (e) {
+      print(e);
+      rethrow;
+    } catch (e, stacktrace) {
+      // the whole stacktrace is heavy
+      print(stacktrace);
+      throw GrpcError.unknown('$e');
     }
   }
 }