+++ packages/fence_service/test/firm_chain_boutiques_test.dart
@@ -74,7 +74,10 @@ void main() async {
 
   test('test upsertOneBoutique', () async {
     final boutiqueLili = chain.boutiques.first..name = 'Lili boutique test';
-    final response = await fenceService.updateOneBoutique(null, boutiqueLili);
+    final response = await fenceService.updateOneBoutique(
+        null,
+        BoutiqueRequest(
+            chainId: boutiqueLili.chainId, boutique: boutiqueLili.boutique));
     expect(response.type, StatusResponse_Type.UPDATED);
     final response2 = await fenceService.readAllChains(null, Empty());
     expect(response2.chains.first.boutiques.first.name, 'Lili boutique test');