+++ packages/protos/protos_weebi/lib/protos_weebi_io.dart
@@ -4,6 +4,10 @@ export 'package:protos_weebi/src/generated/article/article.pb.dart';
 export 'package:protos_weebi/src/generated/article/article.pbenum.dart';
 export 'package:protos_weebi/src/generated/article/article.pbjson.dart';
 
+export 'package:protos_weebi/src/generated/article/photo.pb.dart';
+export 'package:protos_weebi/src/generated/article/photo.pbenum.dart';
+export 'package:protos_weebi/src/generated/article/photo.pbjson.dart';
+
 export 'package:protos_weebi/src/generated/article/article_service.pb.dart';
 export 'package:protos_weebi/src/generated/article/article_service.pbenum.dart';
 export 'package:protos_weebi/src/generated/article/article_service.pbgrpc.dart';
@@ -12,11 +16,6 @@ export 'package:protos_weebi/src/generated/article/article_service.pbjson.dart';
 export 'package:protos_weebi/src/generated/article/category.pb.dart';
 export 'package:protos_weebi/src/generated/article/category.pbenum.dart';
 export 'package:protos_weebi/src/generated/article/category.pbjson.dart';
-
-export 'package:protos_weebi/src/generated/article/category_service.pb.dart';
-export 'package:protos_weebi/src/generated/article/category_service.pbenum.dart';
-export 'package:protos_weebi/src/generated/article/category_service.pbgrpc.dart';
-export 'package:protos_weebi/src/generated/article/category_service.pbjson.dart';
 
 export 'package:protos_weebi/src/generated/boutique.pb.dart';
 export 'package:protos_weebi/src/generated/boutique.pbenum.dart';