+++ packages/fence_service/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.2
+
+- removed authenticateWithDevice
+- removed readUserPermissionsByToken
+
 ## 1.0.1
 - ease reading own user info