+++ packages/protos/protos_weebi/CHANGELOG.md
@@ -1,3 +1,23 @@
+## 1.0.7+2
+
+- removed authenticateWithDevice
+- removed readUserPermissionsByToken
+
+
+## 1.0.7+1
+- addMany for article, contact, ticket
+- handle photo for articles
+- lastTouchTimestampUTC for lighter sync
+- breaking change on articleService ArticleRequest->CalibreRequest for more coherence
+
+## 1.0.6+3
+
+- boutique also needs to be decomposed in BoutiquePb and BoutiqueMongo 
+
+## 1.0.6+2
+
+- keep categories within articleservice
+
 ## 1.0.6+1
 
 - new offline permission: canUpdateContactBalanceOffline